fix: validate DataSet existence, use long in GetGridValue, re-fetch after Halt
- Add DataSet existence check before creating HPO job (prevents FK
  violation producing 500 instead of proper 404)
- Use long arithmetic in GetGridValue to prevent overflow when
  index * intStep exceeds int.MaxValue
- Re-fetch hpoJob after StopHpoJobAsync in Halt to return fresh
  trial statuses in the response

+++ web-api/platypus/platypus/Logic/HpoLogic.cs
@@ -292,7 +292,7 @@ namespace Nssol.Platypus.Logic
                 case "int":
                     int intMin = (int)(param.Min ?? 0);
                     int intStep = (int)(param.Step ?? 1);
-                    return (intMin + index * intStep).ToString();
+                    return ((long)intMin + (long)index * (long)intStep).ToString();
                 case "float":
                     double floatMin = param.Min ?? 0.0;
                     double floatStep = param.Step ?? 0.1;
